+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Rewrote account creation to delegate to account controller's create
   action
+* Extracted authentication strategy into EasyAuth-OAuthCore gem